refactor: 모델 디렉토리 구조 수정
Lotto, Payment 디렉토리 생성
