+++ src/js/Model/Lotto/Lotto.js
