+++ src/js/Model/Lotto/LottoModel.js
